+++ packages/colors/src/scales/index.ts
@@ -1,0 +1,7 @@
+export * from './gray';
+export * from './blue';
+export * from './green';
+export * from './red';
+export * from './yellow';
+export * from './orange';
+export * from './purple';